don't warn on rest_* having requires

+++ package-lists/prepare_patterns.sh
@@ -9,7 +9,9 @@ export EXPLICIT_UNAME=$2
 
 # call out architecture specifics
 for pat in toinstall/rest_*/requires toinstall/rest_*/recommends; do
-  sh $RPM_SOURCE_DIR/preprocess $pat > t && mv t $pat
+  if test -f $pat; then
+    sh $RPM_SOURCE_DIR/preprocess $pat > t && mv t $pat
+  fi
 done
 
 # fill up REST-DVD-SUGGESTS
